Code Separation
This patch creates two classes based around date and quantity text
boxes.  This allows the logic that is specific to each type of text box
input to be split off into a different class than that of MainWindow.
This also means that these UI objects must be populated in code,
but with the added modularization of the code, making it cleaner
to parse, debug, and improve.